+++ src/main/java/crawlercommons/domains/package.html
@@ -16,7 +16,7 @@ limitations under the License.
 -->
 <html>
 <body>
-Classes contained within the url package relate to the definition of Top Level
+Classes contained within the domains package relate to the definition of Top Level
 Domain's, various domain registrars and the effective handling of such domains.
 </body>
 <html>
